Updated deleting of the objects, creating of custom lists

+++ app.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 
 const app = express();
 
@@ -22,6 +23,10 @@ app.use(bodyParser.urlencoded({extended: true}));
 // and put all of your view files there
 app.set("view engine", "ejs");
 
+// Make Mongoose use `findByIdAndRemove()`. Note that this option is `true`
+// by default, you need to set it to false.
+mongoose.set('useFindAndModify', false);
+
 mongoose.connect("mongodb://localhost:27017/todolistDB", {useNewUrlParser: true, useUnifiedTopology: true});
 
 const itemsSchema = mongoose.Schema(
@@ -30,18 +35,29 @@ const itemsSchema = mongoose.Schema(
     }
 );
 
+const listSchema = mongoose.Schema(
+    {
+        listName: String,
+        items: [itemsSchema]
+    }
+);
+
 const Item = mongoose.model("Item", itemsSchema);
+const List = mongoose.model("List", listSchema);
+
+const item1 = new Item({name: "Welcome to your todolist!"});
+const item2 = new Item ({name: "Hit the + button to add a new item."});
+const item3 = new Item ({name: "<-- Hit this to delete an item."});
+
+const defaultItems = [item1, item2, item3];
 
 app.get("/", (req, res) => {
 
     Item.find({}, (err, foundItems)=>{
         if(foundItems.length === 0) {
-            const item1 = new Item({name: "Welcome to your todolist!"});
-            const item2 = new Item ({name: "Hit the + button to add a new item."});
-            const item3 = new Item ({name: "<-- Hit this to delete an item."});
-            Item.insertMany([item1, item2, item3], (err) => {
+            Item.insertMany(defaultItems, (err) => {
                 if(err) console.log(err);
-                else console.log("Succesfully saved default items to the todolistDB");
+                else console.log("Succesfully created and saved default items to the todolistDB");
             }) ;
             res.redirect("/");
         }
@@ -53,34 +69,74 @@ app.get("/", (req, res) => {
 
 app.post("/", (req, res) => {
     const itemName = req.body.task;
-    
-    const item = new Item({name: itemName});
-    item.save();
+    const listName = req.body.list;
 
-    res.redirect("/");
+    const item = new Item({name: itemName});
+
+    if(listName === "Today") {
+        item.save();
+        res.redirect("/");
+    } else {
+        List.findOne({listName: listName}, (err, foundList) => {
+            foundList.items.push(item);
+            foundList.save();
+            res.redirect("/" + listName);
+        });
+    }
 });
 
 app.post("/delete/:_id", (req, res) => {
-    console.log(req.params._id);
     const id = req.params._id;
+    const listName = req.body.listName;
     // if u don't specify callback function, then it will find element by id
     // and return it, but not remove it.
-    Item.findByIdAndRemove(id, (err) => {
-            if(err) console.log("Error while deleting element with _id: " + id + ", error: " + err);
-            else console.log("Succesfully deleted element with _id: " + id);
-        }
-    );
-    res.redirect("/");
+    if(listName === "Today") {
+        Item.findByIdAndRemove(id, (err) => {
+                if(err) console.log("Error while deleting element with _id: " + id + ", error: " + err);
+                else console.log("Succesfully deleted element with _id: " + id);
+                res.redirect("/");
+            }
+        );
+    } else {
+         List.findOneAndUpdate(
+            {listName: listName}, 
+            {$pull: {items: {_id: id}}},
+            (err, foundList) => {
+                if(!err) res.redirect("/" + listName);
+            }
+         );
+    }
 });
 
+app.get("/:customListName", (req, res) => {
+    const customListName = _.capitalize(req.params.customListName);
+    List.findOne({ listName: customListName}, (err, foundList) => {
+        if(!err){
+            if(!foundList) {
+                console.log(customListName + " list does not exists.")
+                // Create a new list
+                const newList = new List({
+                    listName: customListName,
+                    items: defaultItems
+                });
 
-app.get("/work", (req, res) => {
-    const tasks = {
-        title: "Work List",
-        list: workList
-    };    
+                // Insert a new created list with default items to the
+                // "todolistDB", collection "lists"
+                newList.save();
+                res.redirect("/" + customListName);
+            } else {
+                res.render("list", { tasks: {title:foundList.listName, list: foundList.items}});
+            }
+        }    
+    });
 
-    res.render("list", {tasks: tasks});
+    const list = new List({
+        listName: customListName,
+        items: defaultItems
+    });
+
+    list.save();
+
 });
 
 app.get("/about", (req, res) => {
